Porting fixes from Blender.
Fix T37293: trying to fix scrubbing not stopping playback

Fix T39607: Audio not in synch when the blend file loads.

The problem here was that animation buffers got initialized with zeros in the beginning for unknown parts. Now it gets initialized with the first known value.
The bug's result was that the animation of the pitch started with 0 on first playback and thus any seeking while the pitch is zero resulted in seeking to the beginning.

Fix T40280: sequencer sound strips with an end at a negative time kept playing

The bug was caused by using negative numbers as the end for playing forever (or until the end of the sound is reached) in the library. This was used with speaker objects which have an end of FLT_MAX now instead and the negative number interpretation was removed. I hope this doesn't break anything else.

Fix T41133: Audio: Speakers with animated pitch cause clicks in rendered animations, crashes or huge filesizes in rendered audio

Crash happened when the pitch value got <= 0, preventing this now.